refactor(ci): move functions for remote in run_remote.rs and get rid of ssh.rs
