+++ crates/nixci/src/command/mod.rs
@@ -2,3 +2,4 @@
 pub mod core;
 pub mod gh_matrix;
 pub mod run;
+pub mod run_remote;
