+++ crates/nixci/src/nix/mod.rs
@@ -1,5 +1,4 @@
 //! Nix-specific types and functions
 pub mod devour_flake;
 pub mod lock;
-pub mod ssh;
 pub mod system_list;
